+++ src/Plotly.Blazor/Plotly/Marker.cs
@@ -4,7 +4,7 @@
     {
         public object Color { get; set; }
         public object[] Colors { get; set; }
-        public float? Opacity { get; set; }
+        public double? Opacity { get; set; }
         public Line Line { get; set; } = new Line();
         public int? Size { get; set; }
